+++ include/DataPoint.h
@@ -9,13 +9,13 @@ using namespace std;
 class DataPoint {
     private:
         string label;
-        vector<double> data;
+        vector<float> data;
     public:
-        DataPoint(string& label, vector<double>& data);
+        DataPoint(string& label, vector<float>& data);
         string getLabel() const;
         void setLabel(string label);
-        const vector<double>& getData() const;
-        void setData(int i, double val);
+        const vector<float>& getData() const;
+        void setData(int i, float val);
         Distance calcDistance(const DataPoint& x) const;
 };
